requirements.txt
curl_cffi>=0.7.0
requests>=2.31.0
pyyaml>=6.0
pydantic>=2.0
python-dotenv>=1.0.0
